requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
wikipedia>=1.4.0
duckduckgo-search>=4.1.1 